Add a new exportable interface that outputs EGL images
This new interface is intended to make the backend easy to use by
EGL-based applications. A single callback 'export_egl_image' will
provide applications with an already constructed, ready to render
EGL image. A side benefit of this is that the actual protocol used
by the backend's nested compositor (wl_surface, linux-dmabuf, etc)
becomes transparent to the application.

This new interface is built conditionally by setting the
-DEXPORTABLE_EGL cmake option. It is ON by default.