Improve how CMake looks for wayland-scanner
This improves FindWaylandScanner.cmake to allow it to find the
wayland-scanner tool in any of the following ways (in order of
preference):

  1. By passing -DWAYLAND_SCANNER=path/to/wayland-scanner to CMake.
  2. Looking for wayland-scanner in the current $PATH
  3. Figuring out the location from pkg-config (as before).

This will make the package friendlier to cross-compilation, and
then build systems can either prepend their locations for host
tools to $PATH (covered by 2. above), or they can specify the
tool path in the command line (covered by 1. above) if they need
to be more specific.

The motivation for this patch is to avoid kludges like the following:

  https://patchwork.ozlabs.org/comment/2172010/