Handle the case of new target created for the same view backend
The client can remove a target and create a new one for the same view
backend. In that case we need to unregister the current view backend in
the wayland compositor before registering the new one. This patch renames
WS::Instance::createSurface as registerSurface to clarify that it's not
creating a new surface but registering a new one for the given id. Also
ViewBackend::m_id has been renamed as m_surfaceId to clarify that's not
the backend identifier, but the current surface identifier.