Check pending wayland events in the queue on source preapre
Instead of doing it in the check. If the event was read by another
source we will have it in the queue already on prepare. In that case we
return TRUE to proceed to dispatch. This patch also simplifies the error
handling by checking the error events on dispatch and ensures the calls
to prepare_read and read_events/cancel_read are always balanced.