cmake: Pass -Wall and check compiler flags are supported
Adds checks to ensure that command line flags to be passed to the C/C++
compilers are supported before adding them to the list of flags. This
also leaves to CMake the responsibility of determining which compiler
flags are needed to choose C++11/C99, instead of passing -std=xyz,
because different compiler might use other options for this.