cmake: Use an imported library for libwpe
This modifies FindWPE.cmake to make it create the WPE::libwpe imported
library target, which is then passed to target_link_libraries(). The
advantage of this approach is that adding the imported library as a
dependency will automatically propagate the compiler flags and include
directories when building source files, and also propagate all the
needed linker flags and library names for the linker command line.

Note that how find_path() and find_library() invocations are done in
FindWPE.cmake is subtly changed: pkg-config is considered to be the
"source of truth" for how to use a library, but we still want to fall
back to let CMake try and find it in case pkg-config was not able to.
Therefore, the output variables are renamed to make it clearer that
they will contain only one directory (WPE_INCLUDEDIR) and one library
(WPE_LIBRARY):

 - If pkg-config was successful, the include/library directories
   obtained from it are used as hints where to find the files, but the
   actual compiler and linker options are taken from the PkgConfig::WPE
   imported library target.

 - Else, if pkg-config failed to return something there is no imported
   library target defined, but it is still possible that find_path() and
   find_library() manage to find something. In this case the imported
   library target is constructed by hand with the paths found.

Lastly, WPE_INCLUDEDIR and WPE_LIBRARY are used to determine whether
findind the library was successful (they are passed as required to
FIND_PACKAGE_HANDLE_STANDARD_ARGS), because they have to be always
defined to some valid path regardless of whether pkg-config is providing
the information or CMake found the paths by itself.