cmake: Add missing gobject module
This is needed to avoid the linker bailing out due to undefined symbols
when "--no-undefined" is used. The fact that things worked before was by
accident because the FDO backend is always used by binaries which end up
pulling libgobject-2.0.so indirectly at runtime, so the dynamic linker
(at least of GNU/Linux) ended up being able to resolve the symbols. At
any rate, pulling in the linker flags for the gobject module is the
correct thing to do.