view-backend-exportable-fdo-egl: improve buffer wl_resource management in the deprecated implementation
Use a destroy handler on the buffer wl_resource in order to properly track
resource's lifetime. In case the resource was destroyed while corresponding
wl_resource was exported to the client, the subsequent release should avoid
dispatching buffer release notification on the already-destroyed resource.

ClientBundleEGLDeprecated tracks the exported resources (on which EGLImage
objects are based) in a wl_list. Upon resource destruction, the tracked
resource is removed.

When the buffer resource is returned, it is released if it is still tracked
and was not yet destroyed. The corresponding EGLImage is destroyed
unconditionally.