view-backend-exportable-fdo-egl: clean up management of buffer resources in the wpe_fdo_egl_exported_image implementation
The ClientBundleEGL implementat that operates with wpe_fdo_egl_exported_image
objects already relies on destroy listeners to properly track wl_resource
lifetime, so these changes only polish things.

The basics remain the same: if the wl_resource is destroyed, the
wpe_fdo_egl_exported_image object is destroyed immediately if it's not currently
exported, otherwise it's destroyed when the client returns it.