cmake: Use imported targets for Wayland libraries
This modifies FindWayland.cmake to accept a list of components to check
for, and to define imported targets for the Wayland libraries. A target
is defined for each one of the components. This same module can now be
used to check for the client, server, cursor, and EGL Wayland libraries.

Also, FindWaylandEGL.cmake is no longer needed, as "egl" can be passed
as one of the components to FindWayland.cmake; hence its removal.