Make WS::Instance::initialize crash on failure
We need crashes here to debug
https://bugs.webkit.org/show_bug.cgi?id=201507. Otherwise, we're never
going to figure out what is going wrong. I decided to use g_error()
rather than g_warning() because we are going to want to be able to
examine stack and member variables to see as much of what went wrong as
possible. Having a backtrace will be much more useful for debugging than
just having a warning.

But because the function is designed to return false on failure, and
this commit subverts that, we will eventually want to revert this,
either turning the crashes into warnings or just removing the warnings
entirely. I've added a FIXME so we don't forget to do this.