Install to the libwpe backends directory
Use pkg-config to obtain the value of the "backendsdir" variable to use
the value as location for installing the backend. This also changes the
pkg-config .pc file to have that list that directory in the Libs variable,
so applications which need to link into libWPEBackend-fdo-1.0 will still
be able to find it.