Provide EGLStream mode
Add initial support for the EGLStream buffer sharing mechanism that's
most prominently implemented in NVIDIA graphics stacks.

Existing integration of the EGLStream and the Wayland IPC mechanism
allows the client behavior to remain unchanged, meaning the Wayland
EGL platform can still be used. For this to work, we are required to
provide a `wl_eglstream_controller` object on the internal server.
This is done when `wpe_fdo_initialize_eglstream()` is invoked.

The resulting Wayland resources are exported to the user through the
new EGLStream-oriented interface, `wpe_view_backend_exportable_fdo_eglstream_client`.
`export_eglstream_producer_resource()` provides the wl_resource
object from which the user is required to extract the EGLStream
producer and immediately associate it with an EGLStream consumer.
The callback can be invoked multiple times for the same view, e.g.
a new EGLStream is produced for each change in view's size.

The `notify_eglstream_frame()` is a simple callback that signals
whenever a new frame is available on the EGLStream, allowing the user
to put the EGLStream consumer to work.

The new mode doesn't impose additional dependencies, but the
implementation and the related APIs are still subject to change.