CMake: Remove EXPORTABLE_EGL build option
Always build the EGL exportable interface. This is fine as the EGL
dependency is always needed anyway, and the amount of code it adds
is small. Removing the option will not break builds if the option is
passed in the command line: CMake will emit a warning instead telling
that the option has not been declared, and continue ahead configuring
the build.