meson: Only add a wayland-egl dependency if needed
Some drivers provide the wl_egl_* symbols their EGL library (either
directly or indirectly in some other one needed by it) and do not
ship a wayland-egl library. In that case, checking for wayland-egl
fails even if the symbols are usable. To accomodate for this, check
whether wl_egl_window_create is usable before trying to add the
dependency on wayland-egl.