view-backend-exportable-private: use SOCK_CLOEXEC
I noticed that each time WebKit launches a new web process, the new
process has two more sockets open than the previous one. This becomes
severe when many web processes are in use simultaneously. In Epiphany,
I can open roughly 70 tabs before other things start to go wrong,
meaning 140 fds leaked into the final web process.

Turns out we are missing SOCK_CLOEXEC in ViewBackend::initialize, and so
the sockets we create here remain open in the child process. Let's fix
that.

But wait! Shouldn't the client fd be leaked to the web process? That's
why it's created, after all: to allow the UI process to talk to the web
process. So why is it OK to use CLOEXEC? Because the actual fd passed to
the child process is not the one created in ViewBackend::initialize, but
the one created using dup() in ViewBackend::clientFd to copy the one
created here. Since we don't pass O_CLOEXEC to dup(), the duped fd will
still leak to the child process as before. And WebKit is careful to
always close this fd after it launches the web process, so it won't leak
to subsequent processes.

(cherry picked from commit 172c66858840754012ddcf6723c487a9c7c73284)