Avoid generated version header
Use a fixed wpebackend-fdo-version.h header which contains #defines for
the version numbers only, which is the location where the canonical
version is read by the build system, and also used by version.h, which
does not need to be generated anymore.

This avoids needing to use relative #include paths, which continued
being a pitfall even after the recent improvements due to the mixup of
headers with the same name (and relative path!) in the preprocessor
search path e.g. wpe/version.h from libwpe.