view backend: Properly unregister surfaces
When a new surface is registered for an existing ViewBackend, unregister
the old one first, to avoid keeping stale registrations. This makes the
WS::Instance::m_viewBackendMap only contain one entry pointing for any
given ViewBackend instance, instead of growing continuosly.

To ensure that the same code path is taken when the wl_client for a
ViewBackend is destroyed, all the operations to be done are moved into
ViewBackend::clientDestroyNotify(), the callback of the wl_client
destruction listener, then ViewBackend::unregisterSurface() only needs
to call wl_client_destroy(). This way the ~ViewBackend destructor only
needs to call unregisterSurface() as well.

Also, avoid doing any operation in ViewBackend::dispatchFrameCallbacks()
and ViewBackend::releaseBuffer() if the client connection has been lost.