Remove view backend bridge association on connection drops
Arrange for calling ViewBackend::unregisterSurface() also when a Surface
is destroyed. This is a no-op if the nested compositor had the chance
to receive and handle a FdoIPC::UnregisterSurface message before the
surface is destroyed; but allows undoing the bridge association when the
Wayland client connection is closed beforehand. The destruction callback
for the surface is guaranteed to always be called by libwayland, even
on dropped connections, to allow cleaning up reagardless of the fate
of its clients.

Among others, this covers the case in which a WPEWebProcess is abruptly
exited, for example due to a crash or a top-level resource load being
blocked by a content filter.

Co-authored-by: Adrian Perez de Castro <aperez@igalia.com>
Acked-by: Adrian Perez de Castro <aperez@igalia.com>
Signed-off-by: Pablo Saavedra <psaavedra@igalia.com>