Set back the ViewBackend to the previous bridge references
Falling back to the previous bridge Id because the current Surface is
gone. This action prevents inconsistencies by keeping the ViewBackend in
a consistent state pointing to a still valid bridge reference.

An example of a scenario that could generate this kind of inconsistency
is a failed load request due to content filter policies. In that case,
the client gone situation is reached because

Co-authored-by: Adrian Perez de Castro <aperez@igalia.com>
Acked-by: Adrian Perez de Castro <aperez@igalia.com>
Signed-off-by: Pablo Saavedra <psaavedra@igalia.com>