view-backend: dispatch pending frame callbacks after a surface is unregistered
This was causing WebKit to stop painting after provisional navigation
fails with PSON enabled. It happened in the following situation:

1. Normal load, view backend registers surface for bridge 1
2. A frame is requested for bridge 1
3. CORS load starts, a new process is launched and view backend
   registers surface for bridge 2
4. Frame for bridge 1 is rendered, but current bridge is 2 so
   ViewBackend::dispatchFrameCallbacks() does nothing.
5. CORS load fails and the provisional load is destroyed so surface for
   bridge 2 is unregistered
6. No more updates because there's still the frame callback pending for
   surface for bridge 1.

This always happens when WebKit starts a CORS navigation that is
converted to a download, because in that case the provisional load fails
with cancelled error. It was never committed and nothing was rendered.
So, while it's ok to stop painting new frames during the provisional
load to avoid flashing, we should check if there's any pending callback
once the current surface is unregistered.