ws: Save and expose damage buffer information
The wl_surface.damage_buffer request can be used by clients to inform
that a given region (in buffer coordinates) is damaged. This request is
used by eglSwapBuffersWithDamage[EXT|KHR], with one call for each rect.

For more info: https://wayland-book.com/surfaces-in-depth/damaging-surfaces.html

This commit stores this information and exposes it to exported EGL image
clients.

It also bumps the nested compositor interface to version 4, which
provides the damage_buffer support.