CI: Run on Ubuntu 20.04
The 18.04 runners are not available anymore as per
https://github.com/actions/virtual-environments/issues/6002 so use
instead the lowest available version at the moment. While at it,
update to the latest version of the upload-artifact action, as v1
is deprecated; and to Python 3.7 as 3.5 is no longer available.
Lastly, update to Meson 0.55 because that's the version now required
by libwpe.